requirements.txt
Django==1.10
djangorestframework==3.4.0
djangorestframework-jwt==1.11.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
pytz==2018.7
